# Use faster hashmap for mode aggregation