--- a/extension/core_functions/aggregate/holistic/mode.cpp
+++ b/extension/core_functions/aggregate/holistic/mode.cpp
@@ -11,6 +11,9 @@
 #include "duckdb/function/create_sort_key.hpp"
 #include "duckdb/function/aggregate/sort_key_helpers.hpp"
 #include "duckdb/common/algorithm.hpp"
+
+#include "parallel_hashmap/phmap.h"
+
 #include <functional>
 
 // MODE( <expr1> )
@@ -30,8 +33,7 @@ struct ModeAttr {
 
 template <class T>
 struct ModeStandard {
-	using MAP_TYPE = unordered_map<T, ModeAttr>;
-
+	using MAP_TYPE = phmap::flat_hash_map<T, ModeAttr>;
 	static MAP_TYPE *CreateEmpty(ArenaAllocator &) {
 		return new MAP_TYPE();
 	}
@@ -46,7 +48,7 @@ struct ModeStandard {
 };
 
 struct ModeString {
-	using MAP_TYPE = OwningStringMap<ModeAttr>;
+	using MAP_TYPE = OwningStringMap<ModeAttr, phmap::flat_hash_map<string_t, ModeAttr, StringHash, StringEquality>>;
 
 	static MAP_TYPE *CreateEmpty(ArenaAllocator &allocator) {
 		return new MAP_TYPE(allocator);