--- a/extension/core_functions/aggregate/holistic/mode.cpp
+++ b/extension/core_functions/aggregate/holistic/mode.cpp
@@ -48,7 +48,8 @@ struct ModeStandard {
 };
 
 struct ModeString {
-	using MAP_TYPE = OwningStringMap<ModeAttr, duckdb_phmap::flat_hash_map<string_t, ModeAttr, StringHash, StringEquality>>;
+	using MAP_TYPE =
+	    OwningStringMap<ModeAttr, duckdb_phmap::flat_hash_map<string_t, ModeAttr, StringHash, StringEquality>>;
 
 	static MAP_TYPE *CreateEmpty(ArenaAllocator &allocator) {
 		return new MAP_TYPE(allocator);