--- a/scripts/package_build.py
+++ b/scripts/package_build.py
@@ -38,6 +38,7 @@ def third_party_includes():
     includes += [os.path.join('third_party', 'utf8proc', 'include')]
     includes += [os.path.join('third_party', 'yyjson', 'include')]
     includes += [os.path.join('third_party', 'zstd', 'include')]
+    includes += [os.path.join('third_party', 'parallel_hashmap')]
     return includes