This is a pull request implementing https://github.com/duckdb/duckdb/discussions/16569.

There are several high-performance hash map libraries available, but to minimize changes, I chose [`greg7mdp/parallel-hashmap`](https://github.com/greg7mdp/parallel-hashmap), which serves as a drop-in replacement for `std::unordered_map`.

The results of the microbenchmark are as follows:

| name | before | after |
|---|---|---|
|benchmark/micro/aggregate/mode_aggregate_integer.benchmark | 1.209s | 0.521s |
|benchmark/micro/aggregate/mode_aggregate_string.benchmark | 5.732s | 2.287s |

(CPU: Core™ i7-13700KF, RAM: 80GB, OS: Ubuntu 22.04.5)