@Mytherin 2025-03-18
Thanks for the PR!

Appreciate the effort here - and the results are good - but we are generally very careful about pulling in third party libraries because they tend to cause complications. In particular high performance libraries often do a lot of unholy things to gain a few percentage points on benchmarks - which tends to cause issues when porting to different platforms and build systems.

We discussed the `mode` function and think this is best handled using a two-tier aggregation using our own aggregate hash table, i.e. this:

```sql
SELECT grp, MODE(value) FROM t GROUP BY ALL;
```

Can be executed as:

```sql
SELECT grp, MAX_BY(value, count) AS mode
FROM 
    (SELECT grp, value, COUNT(*) AS count
    FROM t
    GROUP BY ALL)
GROUP BY ALL;
```

This is already how distinct aggregates work internally - so this would require generalizing that code to work for different aggregates. It also solves memory issues surrounding the aggregate - as the system knows how to spill our own aggregate hash table to disk. The system does not know how to do this for random other containers (like `std` containers, or this hash map).

It is also faster than the proposed hash table, e.g.:

```sql
CREATE TABLE t AS SELECT i%10000 AS grp, ((i/10000)::integer%1000)::string as value FROM range(50000000) tbl(i);

SELECT grp, MODE(value) FROM t GROUP BY ALL;
```

| unordered_map | phmap::flat_hash_map | SQL-level rewrite |
|---------------|----------------------|-------------------|
| 14.8s         | 0.68s                | 0.27s             |

Given that we plan to solve this in a different manner - merging this would effectively only provide a temporary solution while potentially causing a lot of other issues around distributing DuckDB. As such, I would like to thank you again for the effort - but regrettably we will not be able to merge this PR.