@nyanp 2025-03-20
@Mytherin 

Thank you for your comment!

I completely understand your concerns. In general, libraries that prioritize performance can indeed cause issues in open-source software like DuckDB, which is designed for multi-platform deployment.


However, I still believe that `phmap` is worth considering for several reasons:


- `phmap` is based on Google's well-known [Abseil](https://github.com/abseil/abseil-cpp) implementation and has already been tested on various platforms. The Swiss Table algorithm from Abseil has even been ported to [Rust's standard HashMap](https://github.com/rust-lang/hashbrown) and is widely used. This is one of the main reasons why I chose `phmap` over other libraries that might perform better in benchmarks.
- Replacing `std::unordered_map` or `string_map_t` with `phmap` is not only beneficial for optimizing the `MODE` function but also has the potential to improve the performance of other parts of DuckDB in the future.
- The two-tier aggregation you suggested seems difficult to apply to window functions. Even for the `MODE` function, there may still be cases where a hash map is required on the C++ side.


That said, in my use case, the two-tier aggregation you mentioned is quite appealing! Could you share the roadmap for when this optimization will be introduced?