@Mytherin 2025-03-20
I have no doubt that `phmap` is a great alternative to `std::unordered_map` - however - fundamentally the issue is that the interface designed by `std::unordered_map` is always going to be inefficient compared to our own hash table (which is optimized for bulk insertion/processing) - and always going to be opaque in terms of memory usage.

I think for a standard C++ library it's a great drop-in replacement - but we need more knowledge about the containers we are using. The end goal for us is to get rid of all usages of `std::` containers in performance critical paths - not to add more usage of them. 

I agree for the window function such a rewrite won't work - we'll have to come up with some other solution.

> That said, in my use case, the two-tier aggregation you mentioned is quite appealing! Could you share the roadmap for when this optimization will be introduced?

It's hard to give an exact time-line, but it should be implemented in the coming year. It is of course also possible to do the rewrite yourself at the SQL level as a work-around in the meantime.